Automatically generate all SOMA object URIs on ingestion.
Rather than manually assembling URIs and passing them down through the
ingestion along side the collections they belong to, this change always
uses the collection's creation URL along with the default subdirectory
for the entry to store data. This means that the entire hierarchy will
be automatically built according to the conventional storage layout.

Because we now store the creation URI of TileDB Groups and Arrays when
creating them on TileDB Cloud, there is no need for a manual control
over whether to use relative or absolute URIs; the correct type will be
selected in all cases.

This also means that resumption works in more cases. If you ingest data
to a TileDB group like:

    tiledb://my-org/backend://path/to/my/important-soma-dataset

and then you continue ingesting to the new URI that TileDB creates:

    tiledb://my-org/important-soma-dataset

the ingestion resumption will continue to work and data will be added to
the existing path at `backend://path/to/my/important-soma-dataset`.